fix(): refine yaml linter
