+++ bricks/vs/test.config.js
@@ -1,0 +1,5 @@
+// @ts-check
+/** @type {import("@next-core/test-next").TestNextConfig} */
+export default {
+  transformModulePatterns: ["yaml/"],
+};
